MSVC is *SO* fussy with type conversions...

+++ tests/casting.cpp
@@ -74,7 +74,7 @@ TEST_CASE("arby::Uint::from_float() with positive value") {
 }
 
 TEST_CASE("arby::Uint can be constructed from unsigned types smaller than uintmax_t - uint8_t",) {
-    std::uint8_t input = GENERATE(take(1000, random((std::uint16_t)0x00, (std::uint16_t)0xFF)));
+    std::uint8_t input = (uint16_t)GENERATE(take(1000, random((std::uint16_t)0x00, (std::uint16_t)0xFF)));
 
     arby::Uint object = input;
